requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.4
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
numpy==1.25.2 ; python_version >= '3.10'
opencv-contrib-python==4.8.0.76
pdf2image==1.16.3
pillow==10.0.0
pypdf2==2
pytz==2023.3
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
